CC/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-multipart==0.0.18
passlib==1.7.4
pyzbar==0.1.9
pillow==11.0.0
tensorflow==2.15.0
google-cloud-storage==2.11.0
mysql-connector-python==8.0.29
pydantic[email]==2.10.3
pandas==2.2.3
scikit-learn==1.5.2
apscheduler==3.11.0
sqlalchemy  # ORM untuk database
google-cloud  # SDK untuk Google Cloud
